Revert "Add Single block provider for Fulu  (#10169)"
This reverts commit 2246ea0da0e1f671c3ca91c0b0ec127613a4bdd8.
