+++ beacon/sync/src/main/java/tech/pegasys/teku/beacon/sync/gossip/blocks/RecentBlocksFetchService.java
@@ -26,7 +26,6 @@ import tech.pegasys.teku.infrastructure.subscribers.Subscribers;
 import tech.pegasys.teku.spec.datastructures.attestation.ValidatableAttestation;
 import tech.pegasys.teku.spec.datastructures.blocks.SignedBeaconBlock;
 import tech.pegasys.teku.statetransition.blobs.BlockBlobSidecarsTrackersPool;
-import tech.pegasys.teku.statetransition.datacolumns.DasSamplerBasic;
 import tech.pegasys.teku.statetransition.util.PendingPool;
 
 public class RecentBlocksFetchService
@@ -41,7 +40,6 @@ public class RecentBlocksFetchService
   private final PendingPool<SignedBeaconBlock> pendingBlockPool;
   private final PendingPool<ValidatableAttestation> pendingAttestationsPool;
   private final BlockBlobSidecarsTrackersPool blockBlobSidecarsTrackersPool;
-  private final DasSamplerBasic dasBasicSampler;
   private final FetchTaskFactory fetchTaskFactory;
   private final Subscribers<BlockSubscriber> blockSubscribers = Subscribers.create(true);
 
@@ -50,7 +48,6 @@ public class RecentBlocksFetchService
       final PendingPool<SignedBeaconBlock> pendingBlockPool,
       final PendingPool<ValidatableAttestation> pendingAttestationsPool,
       final BlockBlobSidecarsTrackersPool blockBlobSidecarsTrackersPool,
-      final DasSamplerBasic dasBasicSampler,
       final ForwardSync forwardSync,
       final FetchTaskFactory fetchTaskFactory,
       final int maxConcurrentRequests) {
@@ -59,7 +56,6 @@ public class RecentBlocksFetchService
     this.pendingBlockPool = pendingBlockPool;
     this.pendingAttestationsPool = pendingAttestationsPool;
     this.blockBlobSidecarsTrackersPool = blockBlobSidecarsTrackersPool;
-    this.dasBasicSampler = dasBasicSampler;
     this.fetchTaskFactory = fetchTaskFactory;
   }
 
@@ -68,7 +64,6 @@ public class RecentBlocksFetchService
       final PendingPool<SignedBeaconBlock> pendingBlocksPool,
       final PendingPool<ValidatableAttestation> pendingAttestations,
       final BlockBlobSidecarsTrackersPool blockBlobSidecarsTrackersPool,
-      final DasSamplerBasic dasBasicSampler,
       final ForwardSync forwardSync,
       final FetchTaskFactory fetchTaskFactory) {
     return new RecentBlocksFetchService(
@@ -76,7 +71,6 @@ public class RecentBlocksFetchService
         pendingBlocksPool,
         pendingAttestations,
         blockBlobSidecarsTrackersPool,
-        dasBasicSampler,
         forwardSync,
         fetchTaskFactory,
         MAX_CONCURRENT_REQUESTS);
@@ -110,10 +104,6 @@ public class RecentBlocksFetchService
     }
     if (blockBlobSidecarsTrackersPool.containsBlock(blockRoot)) {
       // We already have this block, waiting for blobs
-      return;
-    }
-    if (dasBasicSampler.containsBlock(blockRoot)) {
-      // We already have this block in DAS sampler
       return;
     }
     final FetchBlockTask task = createTask(blockRoot);
