+++ beacon/sync/src/test/java/tech/pegasys/teku/beacon/sync/gossip/blocks/RecentBlocksFetchServiceTest.java
@@ -43,7 +43,6 @@ import tech.pegasys.teku.spec.datastructures.attestation.ValidatableAttestation;
 import tech.pegasys.teku.spec.datastructures.blocks.SignedBeaconBlock;
 import tech.pegasys.teku.spec.util.DataStructureUtil;
 import tech.pegasys.teku.statetransition.blobs.BlockBlobSidecarsTrackersPool;
-import tech.pegasys.teku.statetransition.datacolumns.DasSamplerBasicImpl;
 import tech.pegasys.teku.statetransition.util.PendingPool;
 
 public class RecentBlocksFetchServiceTest {
@@ -60,7 +59,6 @@ public class RecentBlocksFetchServiceTest {
 
   private final BlockBlobSidecarsTrackersPool blockBlobSidecarsTrackersPool =
       mock(BlockBlobSidecarsTrackersPool.class);
-  private final DasSamplerBasicImpl dasSamplerBasic = mock(DasSamplerBasicImpl.class);
 
   private final FetchTaskFactory fetchTaskFactory = mock(FetchTaskFactory.class);
 
@@ -83,7 +81,6 @@ public class RecentBlocksFetchServiceTest {
             pendingBlockPool,
             pendingAttestationsPool,
             blockBlobSidecarsTrackersPool,
-            dasSamplerBasic,
             forwardSync,
             fetchTaskFactory,
             maxConcurrentRequests);
