+++ beacon/sync/src/testFixtures/java/tech/pegasys/teku/beacon/sync/SyncingNodeManager.java
@@ -63,7 +63,6 @@ import tech.pegasys.teku.statetransition.block.BlockImportChannel;
 import tech.pegasys.teku.statetransition.block.BlockImporter;
 import tech.pegasys.teku.statetransition.block.BlockManager;
 import tech.pegasys.teku.statetransition.block.ReceivedBlockEventsChannel;
-import tech.pegasys.teku.statetransition.datacolumns.DasSamplerBasic;
 import tech.pegasys.teku.statetransition.forkchoice.ForkChoice;
 import tech.pegasys.teku.statetransition.forkchoice.MergeTransitionBlockValidator;
 import tech.pegasys.teku.statetransition.forkchoice.NoopForkChoiceNotifier;
@@ -215,7 +214,6 @@ public class SyncingNodeManager {
             pendingBlocks,
             pendingAttestations,
             BlockBlobSidecarsTrackersPool.NOOP,
-            DasSamplerBasic.NOOP,
             syncService,
             fetchBlockTaskFactory);
     recentBlocksFetcher.subscribeBlockFetched(blockManager::importBlock);
