+++ ethereum/statetransition/src/main/java/tech/pegasys/teku/statetransition/datacolumns/DataColumnSamplingTracker.java
@@ -13,21 +13,15 @@
 
 package tech.pegasys.teku.statetransition.datacolumns;
 
-import static com.google.common.base.Preconditions.checkArgument;
-
 import java.util.List;
-import java.util.Optional;
 import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicBoolean;
-import java.util.concurrent.atomic.AtomicReference;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.apache.tuweni.bytes.Bytes32;
 import tech.pegasys.teku.infrastructure.async.SafeFuture;
 import tech.pegasys.teku.infrastructure.unsigned.UInt64;
-import tech.pegasys.teku.spec.datastructures.blocks.SignedBeaconBlock;
-import tech.pegasys.teku.spec.datastructures.blocks.SlotAndBlockRoot;
 import tech.pegasys.teku.spec.datastructures.util.DataColumnSlotAndIdentifier;
 import tech.pegasys.teku.statetransition.blobs.RemoteOrigin;
 
@@ -37,8 +31,7 @@ record DataColumnSamplingTracker(
     List<UInt64> samplingRequirement,
     Set<UInt64> missingColumns,
     AtomicBoolean rpcFetchScheduled,
-    SafeFuture<List<UInt64>> completionFuture,
-    AtomicReference<Optional<SignedBeaconBlock>> block) {
+    SafeFuture<List<UInt64>> completionFuture) {
   private static final Logger LOG = LogManager.getLogger();
 
   static DataColumnSamplingTracker create(
@@ -54,25 +47,7 @@ record DataColumnSamplingTracker(
         samplingRequirement,
         missingColumns,
         new AtomicBoolean(false),
-        new SafeFuture<>(),
-        new AtomicReference<>(Optional.empty()));
-  }
-
-  public Optional<SignedBeaconBlock> getBlock() {
-    return block.get();
-  }
-
-  public boolean setBlock(final SignedBeaconBlock block) {
-    final SlotAndBlockRoot slotAndBlockRoot = new SlotAndBlockRoot(slot, blockRoot);
-    checkArgument(block.getSlotAndBlockRoot().equals(slotAndBlockRoot), "Wrong block");
-    final Optional<SignedBeaconBlock> oldBlock = this.block.getAndSet(Optional.of(block));
-    if (oldBlock.isPresent()) {
-      return false;
-    }
-
-    LOG.debug("Block received for {}", slotAndBlockRoot::toLogString);
-
-    return true;
+        new SafeFuture<>());
   }
 
   boolean add(final DataColumnSlotAndIdentifier columnIdentifier, final RemoteOrigin origin) {
