+++ ethereum/statetransition/src/main/java/tech/pegasys/teku/statetransition/forkchoice/DataColumnSidecarAvailabilityChecker.java
@@ -60,7 +60,7 @@ public class DataColumnSidecarAvailabilityChecker implements AvailabilityChecker
       }
       default -> {
         dataAvailabilitySampler
-            .checkDataAvailability(block)
+            .checkDataAvailability(block.getSlot(), block.getRoot())
             .finish(
                 sampleIndices -> {
                   validationResult.complete(DataAndValidationResult.validResult(sampleIndices));
