+++ ethereum/statetransition/src/test/java/tech/pegasys/teku/statetransition/datacolumns/DasPreSamplerTest.java
@@ -82,7 +82,8 @@ public class DasPreSamplerTest {
 
     // Setup for the block that will be sampled
     when(custodyGroupCountManager.getSamplingColumnIndices()).thenReturn(List.of());
-    when(sampler.checkDataAvailability(blockToSample)).thenReturn(SafeFuture.completedFuture(null));
+    when(sampler.checkDataAvailability(blockToSample.getSlot(), blockToSample.getRoot()))
+        .thenReturn(SafeFuture.completedFuture(null));
 
     dasPreSampler.onNewPreImportBlocks(List.of(blockToSample, blockToSkip));
 
@@ -91,10 +92,10 @@ public class DasPreSamplerTest {
     verify(sampler).checkSamplingEligibility(blockToSkip.getMessage());
 
     // Verify processing for the required block
-    verify(sampler).checkDataAvailability(blockToSample);
+    verify(sampler).checkDataAvailability(blockToSample.getSlot(), blockToSample.getRoot());
 
     // Verify no processing for the skipped block
-    verify(sampler, never()).checkDataAvailability(eq(blockToSkip));
+    verify(sampler, never()).checkDataAvailability(eq(blockToSkip.getSlot()), any());
 
     verify(sampler).flush();
   }
@@ -111,7 +112,8 @@ public class DasPreSamplerTest {
     when(sampler.checkSamplingEligibility(block.getMessage())).thenReturn(REQUIRED);
     when(custodyGroupCountManager.getSamplingColumnIndices()).thenReturn(columnIndices);
     when(custody.hasCustodyDataColumnSidecar(any())).thenReturn(SafeFuture.completedFuture(false));
-    when(sampler.checkDataAvailability(block)).thenReturn(SafeFuture.completedFuture(null));
+    when(sampler.checkDataAvailability(block.getSlot(), block.getRoot()))
+        .thenReturn(SafeFuture.completedFuture(null));
 
     dasPreSampler.onNewPreImportBlocks(List.of(block));
 
@@ -119,7 +121,7 @@ public class DasPreSamplerTest {
     verify(custody).hasCustodyDataColumnSidecar(col5);
     verify(sampler, never())
         .onNewValidatedDataColumnSidecar(any(DataColumnSlotAndIdentifier.class), any());
-    verify(sampler).checkDataAvailability(block);
+    verify(sampler).checkDataAvailability(block.getSlot(), block.getRoot());
     verify(sampler).flush();
   }
 
@@ -139,7 +141,8 @@ public class DasPreSamplerTest {
     when(custody.hasCustodyDataColumnSidecar(col2)).thenReturn(SafeFuture.completedFuture(true));
     when(custody.hasCustodyDataColumnSidecar(col4)).thenReturn(SafeFuture.completedFuture(false));
 
-    when(sampler.checkDataAvailability(block)).thenReturn(SafeFuture.completedFuture(null));
+    when(sampler.checkDataAvailability(block.getSlot(), block.getRoot()))
+        .thenReturn(SafeFuture.completedFuture(null));
 
     dasPreSampler.onNewPreImportBlocks(List.of(block));
 
@@ -151,7 +154,7 @@ public class DasPreSamplerTest {
                 id -> id.columnIndex().equals(UInt64.valueOf(4))),
             any());
 
-    verify(sampler).checkDataAvailability(block);
+    verify(sampler).checkDataAvailability(block.getSlot(), block.getRoot());
     verify(sampler).flush();
   }
 }
