+++ ethereum/statetransition/src/test/java/tech/pegasys/teku/statetransition/datacolumns/DasSamplerBasicTest.java
@@ -35,7 +35,6 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import tech.pegasys.teku.infrastructure.async.SafeFuture;
 import tech.pegasys.teku.infrastructure.async.StubAsyncRunner;
-import tech.pegasys.teku.infrastructure.metrics.StubMetricsSystem;
 import tech.pegasys.teku.infrastructure.time.StubTimeProvider;
 import tech.pegasys.teku.infrastructure.unsigned.UInt64;
 import tech.pegasys.teku.spec.Spec;
@@ -52,7 +51,7 @@ import tech.pegasys.teku.statetransition.datacolumns.retriever.DataColumnSidecar
 import tech.pegasys.teku.statetransition.util.RPCFetchDelayProvider;
 import tech.pegasys.teku.storage.client.RecentChainData;
 
-public class DasSamplerBasicImplTest {
+public class DasSamplerBasicTest {
   private static final Spec SPEC = TestSpecFactory.createMinimalFulu();
   private static final List<UInt64> SAMPLING_INDICES =
       List.of(UInt64.valueOf(5), UInt64.ZERO, UInt64.valueOf(2));
@@ -60,7 +59,6 @@ public class DasSamplerBasicImplTest {
   private final StubTimeProvider stubTimeProvider = StubTimeProvider.withTimeInMillis(0);
   private final StubAsyncRunner asyncRunner = new StubAsyncRunner(stubTimeProvider);
   private final RPCFetchDelayProvider rpcFetchDelayProvider = mock(RPCFetchDelayProvider.class);
-  private final StubMetricsSystem metricsSystem = new StubMetricsSystem();
 
   private RecentChainData recentChainData;
   private DataColumnSidecarCustody custody;
@@ -68,7 +66,7 @@ public class DasSamplerBasicImplTest {
   private CurrentSlotProvider currentSlotProvider;
   private final DataStructureUtil dataStructureUtil = new DataStructureUtil(0, SPEC);
 
-  private DasSamplerBasicImpl sampler;
+  private DasSamplerBasic sampler;
 
   @BeforeEach
   public void setUp() {
@@ -86,7 +84,7 @@ public class DasSamplerBasicImplTest {
     when(custody.onNewValidatedDataColumnSidecar(any(), any())).thenReturn(SafeFuture.COMPLETE);
 
     sampler =
-        new DasSamplerBasicImpl(
+        new DasSamplerBasic(
             SPEC,
             asyncRunner,
             currentSlotProvider,
@@ -94,9 +92,7 @@ public class DasSamplerBasicImplTest {
             custody,
             retriever,
             custodyGroupCountManager,
-            recentChainData,
-            metricsSystem,
-            64);
+            recentChainData);
   }
 
   @Test
@@ -147,13 +143,13 @@ public class DasSamplerBasicImplTest {
 
   @Test
   void checkDataAvailability_shouldAddToTrackerAndReturnCompletionFuture() {
-    final SignedBeaconBlock block = dataStructureUtil.randomSignedBeaconBlock();
     final SlotAndBlockRoot slotAndBlockRoot =
-        new SlotAndBlockRoot(block.getSlot(), block.getRoot());
+        new SlotAndBlockRoot(dataStructureUtil.randomSlot(), dataStructureUtil.randomBytes32());
 
     when(retriever.retrieve(any())).thenReturn(new SafeFuture<>());
 
-    final SafeFuture<List<UInt64>> completionFuture = sampler.checkDataAvailability(block);
+    final SafeFuture<List<UInt64>> completionFuture =
+        sampler.checkDataAvailability(slotAndBlockRoot.getSlot(), slotAndBlockRoot.getBlockRoot());
 
     assertSamplerTracker(
         slotAndBlockRoot.getBlockRoot(), slotAndBlockRoot.getSlot(), SAMPLING_INDICES);
@@ -170,14 +166,13 @@ public class DasSamplerBasicImplTest {
   @Test
   @SuppressWarnings("FutureReturnValueIgnored")
   void checkDataAvailability_shouldCallRetrieverForMissingColumnsAndCallCustodyOnRetrieval() {
-    final SignedBeaconBlock block = dataStructureUtil.randomSignedBeaconBlock();
-    final SignedBeaconBlockHeader blockHeader = block.asHeader();
+    final SignedBeaconBlockHeader block = dataStructureUtil.randomSignedBeaconBlockHeader();
     final SlotAndBlockRoot slotAndBlockRoot =
         new SlotAndBlockRoot(block.getMessage().getSlot(), block.getMessage().getRoot());
 
     final Map<UInt64, DataColumnSidecar> missingColumnSidecars =
         SAMPLING_INDICES.stream()
-            .map(index -> dataStructureUtil.randomDataColumnSidecar(blockHeader, index))
+            .map(index -> dataStructureUtil.randomDataColumnSidecar(block, index))
             .collect(
                 Collectors.toUnmodifiableMap(DataColumnSidecar::getIndex, Function.identity()));
     final Map<UInt64, SafeFuture<DataColumnSidecar>> futureSidecarsByIndex =
@@ -204,7 +199,7 @@ public class DasSamplerBasicImplTest {
         .thenReturn(Duration.ofSeconds(1));
 
     // da check is requested
-    sampler.checkDataAvailability(block);
+    sampler.checkDataAvailability(slotAndBlockRoot.getSlot(), slotAndBlockRoot.getBlockRoot());
 
     assertSamplerTrackerHasRPCFetchScheduled(slotAndBlockRoot.getBlockRoot(), true);
 
@@ -238,14 +233,14 @@ public class DasSamplerBasicImplTest {
   @Test
   @SuppressWarnings("FutureReturnValueIgnored")
   void checkDataAvailability_shouldSetRPCFetchedToFalseOnFailureAndRPCFetchesAgain() {
-    final SignedBeaconBlock block = dataStructureUtil.randomSignedBeaconBlock();
     final SlotAndBlockRoot slotAndBlockRoot =
-        new SlotAndBlockRoot(block.getMessage().getSlot(), block.getMessage().getRoot());
+        new SlotAndBlockRoot(dataStructureUtil.randomSlot(), dataStructureUtil.randomBytes32());
 
     when(retriever.retrieve(any()))
         .thenReturn(SafeFuture.failedFuture(new RuntimeException("error")));
 
-    final SafeFuture<List<UInt64>> completionFuture = sampler.checkDataAvailability(block);
+    final SafeFuture<List<UInt64>> completionFuture =
+        sampler.checkDataAvailability(slotAndBlockRoot.getSlot(), slotAndBlockRoot.getBlockRoot());
 
     assertSamplerTracker(
         slotAndBlockRoot.getBlockRoot(), slotAndBlockRoot.getSlot(), SAMPLING_INDICES);
@@ -262,7 +257,8 @@ public class DasSamplerBasicImplTest {
 
     when(retriever.retrieve(any())).thenReturn(new SafeFuture<>());
 
-    final SafeFuture<List<UInt64>> secondCompletionFuture = sampler.checkDataAvailability(block);
+    final SafeFuture<List<UInt64>> secondCompletionFuture =
+        sampler.checkDataAvailability(slotAndBlockRoot.getSlot(), slotAndBlockRoot.getBlockRoot());
 
     assertSamplerTracker(
         slotAndBlockRoot.getBlockRoot(), slotAndBlockRoot.getSlot(), SAMPLING_INDICES);
@@ -276,15 +272,14 @@ public class DasSamplerBasicImplTest {
   @Test
   @SuppressWarnings("FutureReturnValueIgnored")
   void checkDataAvailability_shouldRPCFetchImmediatelyIfNotPreviouslyScheduled() {
-    final SignedBeaconBlock block = dataStructureUtil.randomSignedBeaconBlock();
     final SlotAndBlockRoot slotAndBlockRoot =
-        new SlotAndBlockRoot(block.getMessage().getSlot(), block.getMessage().getRoot());
+        new SlotAndBlockRoot(dataStructureUtil.randomSlot(), dataStructureUtil.randomBytes32());
 
     when(retriever.retrieve(any())).thenReturn(new SafeFuture<>());
 
     when(rpcFetchDelayProvider.calculate(slotAndBlockRoot.getSlot())).thenReturn(Duration.ZERO);
 
-    sampler.checkDataAvailability(block);
+    sampler.checkDataAvailability(slotAndBlockRoot.getSlot(), slotAndBlockRoot.getBlockRoot());
 
     assertRPCFetchInMillis(
         slotAndBlockRoot.getSlot(), slotAndBlockRoot.getBlockRoot(), SAMPLING_INDICES, 0);
@@ -301,16 +296,17 @@ public class DasSamplerBasicImplTest {
     final DataColumnSidecar source2 =
         dataStructureUtil.randomDataColumnSidecar(
             dataStructureUtil.randomSignedBeaconBlockHeader(), dataStructureUtil.randomUInt64());
-    final SignedBeaconBlock block = dataStructureUtil.randomSignedBeaconBlock();
+    final SlotAndBlockRoot source3 =
+        new SlotAndBlockRoot(dataStructureUtil.randomSlot(), dataStructureUtil.randomBytes32());
 
     sampler.onNewValidatedDataColumnSidecar(source1, RemoteOrigin.CUSTODY);
     sampler.onNewValidatedDataColumnSidecar(source2, RemoteOrigin.RPC);
-    sampler.checkDataAvailability(block);
+    sampler.checkDataAvailability(source3.getSlot(), source3.getBlockRoot());
 
     assertThat(sampler.getRecentlySampledColumnsByRoot())
         .containsKey(source1.blockRoot())
         .containsKey(source2.getBeaconBlockRoot())
-        .containsKey(block.getRoot());
+        .containsKey(source3.getBlockRoot());
   }
 
   @Test
