+++ teku/src/main/java/tech/pegasys/teku/cli/options/P2POptions.java
@@ -601,16 +601,6 @@ public class P2POptions {
       arity = "1")
   private int historicalDataMaxQueryQueueSize = P2PConfig.DEFAULT_HISTORICAL_MAX_QUERY_QUEUE_SIZE;
 
-  @Option(
-      names = {"--Xrecently-sampled-blocks-limit"},
-      hidden = true,
-      paramLabel = "<NUMBER>",
-      description =
-          "Maximum number of recently sampled blocks to keep in cache of the DAS sampler.",
-      showDefaultValue = Visibility.ALWAYS,
-      arity = "1")
-  private int maxRecentlySampledBlocks = SyncConfig.DEFAULT_MAX_RECENTLY_SAMPLED_BLOCKS;
-
   private OptionalInt getP2pLowerBound() {
     if (p2pUpperBound.isPresent() && p2pLowerBound.isPresent()) {
       return p2pLowerBound.getAsInt() < p2pUpperBound.getAsInt() ? p2pLowerBound : p2pUpperBound;
@@ -786,8 +776,7 @@ public class P2POptions {
                     .forwardSyncMaxBlobSidecarsPerMinute(forwardSyncBlobSidecarsRateLimit)
                     .forwardSyncBatchSize(forwardSyncBatchSize)
                     .forwardSyncMaxPendingBatches(forwardSyncMaxPendingBatches)
-                    .forwardSyncMaxDistanceFromHead(forwardSyncMaxDistanceFromHead)
-                    .maxRecentlySampledBlocks(maxRecentlySampledBlocks));
+                    .forwardSyncMaxDistanceFromHead(forwardSyncMaxDistanceFromHead));
 
     if (subscribeAllSubnetsEnabled) {
       builder
